requirements.txt
django-mptt
djangorestframework
Django==3.0.7
